add new homepage, change oldhome page to loginpage, updata routing, fix navigating route on authpage
