+++ src/App.jsx
@@ -3,12 +3,13 @@ import "./App.css";
 import Navbar from "./components/Navbar";
 import About from "./pages/About";
 import UserList from "./pages/UserList";
-import Home from "./pages/Home";
+import Login from "./pages/Login";
 import Register from "./pages/RegisterPage";
 import UserProfile from "./pages/UserProfile";
 import Footer from "./components/Footer";
 import Auth from "./pages/Auth";
 import ForgetPassword from "./pages/ForgetPassword";
+import Home from "./pages/Home";
 
 function App() {
   return (
@@ -16,7 +17,8 @@ function App() {
       <Navbar />
 
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<Login />} />
+        <Route path="/home" element={<Home />} />
         <Route path="/userlist" element={<UserList />} />
         <Route path="/about" element={<About />} />
         <Route path="/register" element={<Register />} />
